+++ CHANGELOG.md
@@ -1,5 +1,6 @@
 ## NEXT
 * Fixing bug with `SignalExt::switch_signal_vec`.
+* Adding in `SignalExt::map_value_signal` method.
 
 ## 0.3.31 - (2022-09-10)
 * Adding in serde `Serialize` and `Deserialize` for `MapDiff`.